LPS-72812 Moves test to Asset module
